# apps/backend/openapi.yaml
openapi: "3.0.3"
info:
  title: ArogyaAI Backend API
  version: "1.0.0"
  description: >
    Production-grade OpenAPI specification for ArogyaAI — a health-tech platform.
    Covers authentication, user management, health profiles, data ingestion,
    medical records, AI analysis, RAG-powered explanations, and alert management.
  contact:
    name: ArogyaAI Engineering
    email: engineering@arogyaai.com
  license:
    name: Private
servers:
  - url: http://localhost:8000
    description: Local Development
  - url: https://api.arogyaai.com
    description: Production

# ============================================================
# SECURITY SCHEME
# ============================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token issued at login or refresh.

  # ============================================================
  # REUSABLE SCHEMAS
  # ============================================================
  schemas:
...
    ErrorDetail:
      type: object
      properties:
        detail:
          type: string
          example: "Validation error or descriptive message"

    # --- Auth ---
    SignupRequest:
      type: object
      required: [email, password, full_name]
      properties:
        email:
          type: string
          format: email
          example: "john@arogyaai.com"
        password:
          type: string
          format: password
          minLength: 8
          example: "Secure@1234"
        full_name:
          type: string
          example: "John Doe"
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: "john@arogyaai.com"
        password:
          type: string
          format: password
          example: "Secure@1234"
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIs..."
        refresh_token:
          type: string
          example: "dGhpcyBpcyBhIHJlZnJlc2..."
        token_type:
          type: string
          example: "bearer"
    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          example: "dGhpcyBpcyBhIHJlZnJlc2..."

    # --- User ---
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
          example:
            detail: "Not authenticated"
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
          example:
            detail: "Field 'email' is required"
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
          example:
            detail: "Internal server error"
...
security:
  - BearerAuth: []

# ============================================================
# PATHS / ENDPOINTS
# ============================================================
paths:
...
  /auth/signup:
    post:
      tags: [Authentication]
      summary: Register a new user account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/login:
    post:
      tags: [Authentication]
      summary: Deprecated legacy backend login
      deprecated: true
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '410':
          description: Deprecated. Use Supabase Auth on the frontend and pass the Supabase bearer token to backend APIs.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/logout:
    post:
      tags: [Authentication]
      summary: Revoke current session's refresh token
      responses:
        '200':
          description: Logged out successfully
          content:
            application/json:
              example:
                message: "Session revoked"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/refresh-token:
    post:
      tags: [Authentication]
      summary: Deprecated legacy backend refresh
      deprecated: true
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '410':
          description: Deprecated. Refresh the Supabase session in the frontend client instead.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '500':
          $ref: '#/components/responses/InternalError'
...